+++ tests/test_resources.py
@@ -8,18 +8,18 @@ import os
 import dfetch.resources
 
 
-def test_schema_path() -> None:
-    """Test that schema path can be used as context manager."""
+def test_template_path() -> None:
+    """Test that template path can be used as context manager."""
 
-    with dfetch.resources.schema_path() as schema_path:
-        assert os.path.isfile(schema_path)
+    with dfetch.resources.template_path() as template_path:
+        assert os.path.isfile(template_path)
 
 
-def test_call_schema_path_twice() -> None:
+def test_call_template_path_twice() -> None:
     """Had a lot of problems with calling contextmanager twice."""
 
-    with dfetch.resources.schema_path() as schema_path:
-        assert os.path.isfile(schema_path)
+    with dfetch.resources.template_path() as template_path:
+        assert os.path.isfile(template_path)
 
-    with dfetch.resources.schema_path() as schema_path:
-        assert os.path.isfile(schema_path)
+    with dfetch.resources.template_path() as template_path:
+        assert os.path.isfile(template_path)